Allowing general rectangular maps to be modelled.
You can now upload a map and its latitude and longitude values without the map having to have the shape nx2n.
Would be good to add the ability to set nlat and nlon.

Also fixed a bug related to solving for the planet's orbital period when none is given.