Speeding up code for circ orbits.
Saving a lot of duplicated calculations for circular orbits (e.g. Fin which is always the same, Fstar which is always the same). Can go further and precompute the SSP weighting since that always stays the same as well - should do that next. Also, 60 percent of runtime is computing cp for bell2018 planets, and 18 percent of runtime is computing sigma*T**4 - these are both very smooth and should be easy to use a lookup table for. Using a simple numpy table with 1 Kelvin steps cuts 99% of the runtime for Fout and still gives Fout to within about 0.05 percent.

Results checked against tutorial and are still visually consistent. Runtime does not appear to be negatively affected by changes for eccentric orbits.