Huge speed improvements using lookup tables.
Massively speeding up code (especially for circular orbits) by precomputing flux and cp (for bell2018 planets) as a function of temperature. For a KELT-9b planet, the effects on the phasecurve is only 0.02 ppm peak-to-trough. This has been set to default as the effects should be fairly minimal and the speed-up is huge.